Refactor: Added new package 'menu' - Categorised all menu screens under package 'menu' -Additionally added javadocs to menu classes
